Move events to shared location
